+++ include/Gaffer/PlugAlgo.h
@@ -72,6 +72,11 @@ GAFFER_API bool dependsOnCompute( const ValuePlug *plug );
 template<typename Predicate>
 std::invoke_result_t<Predicate, Plug *> findDestination( Plug *plug, Predicate &&predicate );
 
+/// Visits the plug and its inputs, returing the first `predicate( plug )` result which
+/// evaluates to `true`.
+template<typename Predicate>
+std::invoke_result_t<Predicate, Plug *> findSource( Plug *plug, Predicate &&predicate );
+
 /// Conversion to and from `IECore::Data`
 /// =====================================